requirements.txt
numpy==1.21.0
opencv-contrib-python-headless==4.5.2.54
lensfunpy==1.10.0
imageio==2.16.1
Pillow==9.0.1

pip~=22.0.4
wheel~=0.36.2
setuptools~=57.0.0
shiboken2~=5.15.2.1